pyproject.toml
[project]
name = "gauss-code"
version = "0.1.0"
description = "A Python project managed with uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
